ensure plugins draw correctly even when they have been given focus.
The initial attempt at fixing this bug was to modify the clip, but
this was not correct and only fixed cases where the plugin was on
the left edge of the screen. This CL properly adjusts the plugin's
boundaries to make sure the content is drawn in the appropriate
position.

bug: 3477581
Change-Id: I5958228a93fa5600cd7e24ed9e7e1e26f6b241bc