Fixed 3469204: Crash using downloaded fonts
This change is still pending upstream, but is being applied here now to
make it into M1 before lockdown. Track the WebKit change at:

    https://bugs.webkit.org/show_bug.cgi?id=54758

When font downloading completes, CSSFontFaceSource::fontLoaded() and
CSSSegmentedFontFace::fontLoaded() called pruneTable() to reclaim
GlyphPageTreeNode memory. On Android, downloading and rendering are on
different threads and these GlyphPageTreeNodes may still be in use on
the UI thread.

This change defers freeing those GlyphPageTreeNodes until
CSSFontFaceSource and CSSSegmentedFontFace are destroyed, which occurs
on the UI thread in response to Document::scheduleForcedStyleRecalc(),
which was called by CSSFontSelector::fontLoaded().

Change-Id: Ica74e5be92c00b8c761f5d29a2310c46e62fd9eb