allow multiple invals to return first difference
The shape of the text selection changed recently, so the
inval computation for the shape area was slightly incorrect.

But most of the problem is caused by multiple calls to
draw the selection area. The old assumption was that
successive calls to compute the inval could each work
from the prior state, since only the difference between
the last inval and the current inval needs to be redrawn.

Now, there are multiple calls to compute the area, and any
of them need to return the last difference computed, only
computing a new difference if the selection changed.

Keep track of the last changed selection, so that the last
position, the current position, and the last drawn position
can be tracked correctly.

bug:4073219
Change-Id: Ic9b9d037329e8f792b5ec09c112e665dbdfd0b81