Custom font was appearing all white in Books
This is a strategic rollback of a previous change:
    https://android-git.corp.google.com/g/#change,99977

Although that change fixes a browser crash, it introduces a problematic
UI situation in Books.

The browser crash is rare, but keeps some web pages from being usable.
Nonetheless, upon failure, the browser will shutdown and can be
relaunched for other sites. Another potential point of failure for this
crash has been identified, with a solution to follow in a separate CL.

From discussion with Bart it was decided that the rare browser crash is
less problematic than the usability problem incurred with a book
containing blank pages, and so, regardless of whether or not the
resulting browser crash is fixed, this change needs to be rolled back.

With any luck, however, both issues will be fixed.
(See also http://b/3469204.)

Bug: 4064041
Change-Id: I3416f9e9ae728a99dd84d3df8b0f8d0c0876dbad