More improvements to video player controls.
Bug:4080127

Inset the track by half the width of the thumb so the thumb
travels along the entire length of it.

Make the track narrower vertically, and align it with the
center of the thumb drawing.

Draw the portion of the track that has already been covered
with a different asset.

Change-Id: I37acedc8e8b75bab59fcac7e075c59d70550795c