DO NOT MERGE
cherry pick of https://android-git.corp.google.com/g/#change,106682

b/3491968 stop the hanging of dunkindonuts.com

Sometimes, a sync load can wait forever and lock up the network thread,
here we replace Wait() with TimedWait() with multiple tries to avoid locking.
Unfortunately, TimedWait() doesn't return anything and can't tell if it's
a timeout or a quick pass, so we query timeofday and allow a few seconds for
the system timing error.

Change-Id: I991617130f93b44518d28926725f261d4569d73c