Fix crash in font-handling code caused by repainting while a style recalculation was pending.
The Android FrameCache was being updated while a style recalculation
was pending.  This would cause the cached fonts in the
FontFallbackList to be accessed, but that cache is not necessarily
consistant while a recalculation is pending.
A similar issue was occurring with recordPictureSet.

In updateFrameCache() and recordPictureSet(), early-abort if the document
is waiting for a style recalculation.  In notifyProgressFinished(), remove
the call to updateFrameCache().

BUG=4292199

Change-Id: I2ae80ddc2d596e2f15455f54f4de7dde6e42ade8