DO NOT MERGE
Cherry-pick change I2ae80ddc from master:

Fix crash in font-handling code caused by repainting while a style recalculation was pending.

The Android FrameCache was being updated while a style recalculation
was pending.  This would cause the cached fonts in the
FontFallbackList to be accessed, but that cache is not necessarily
consistant while a recalculation is pending.
A similar issue was occurring with recordPictureSet.

In updateFrameCache() and recordPictureSet(), early-abort if the document
is waiting for a style recalculation.  In notifyProgressFinished(), remove
the call to updateFrameCache().

Bug: 4292199
Bug: 4291311
Change-Id: Ie8bc4cb0637ccb7aee37597ac08fd8abffc149cd