Always check weak global references before using them
We hold weak references to Java objects from native code in several
places to avoid circular reference problems. These objects may become
weakly reachable at any time, after which the GC could null our weak
reference, so we have to null-check at every use.

Note that weak references are nulled before the referent is finalized,
so we can't rely on doing work in the finalizer to wait for the
currently executing message to complete and to remove other messages
from the queue.

This effectively reverts
https://android-git.corp.google.com/g/#change,30955

Bug: 4336862
Change-Id: I431fcac11220cb406c26e31aacb9bda7ea22776e