Merge WebKit at r80534: Fix uses of rounding hacks.
Font rounding hacks have been removed upstream, so there
is no need for our local code anymore. Remove all uses
of the ANDROID_DISABLE_ROUNDING_HACKS macro.

Note that this brings our behavior in line with upstream -
Android will see a change in behavior to before the merge
as we did not disable rounding hacks in all cases.

- SVGInlineTextBox::SVGInlineTextBox() - we don't pass a value for
disableRoundingHacks, so we don't disable hacks, so this behaviour will
change as rounding hacks are disabled now.
- RenderText::createTextBox() - We pass a value for
  disableRoundingHacks, true if any parent is a text control.
  So we will change behavior in any case where the parent is not
  a text control.

Note that although we may see a change in behavior, that change is
consistent with the upstream codebase.

See http://trac.webkit.org/changeset/78846

Change-Id: I5d4d7298083acc519199c4c38ab5acdef5bdc537