Support the screen shot while the video is paused.
Basically, the GL textures generated for the surface texture will be used
for showing the screenshot.
Surface texture will be recycled every time a new video starts.
But GL textures will be recycled either when running out of memory bound,
or when the mediaplayer is deleted.

1. Add the VideoLayerManager as a static instance to manage the info of textures
   allocated for the screenshot. The basic info will be added into the manager
   while the video is played and prepared. During the draw time, the matirx info
   will be updated, too.

2. When there are too many screenshots are created, then we will discard the
   oldest one and just show the grey static rectangle to replace the screenshot.

The corresponding framework change is 112501.

Change-Id: Ifea60c96532500f9c93062cc97f7c4ef978046b0