Remove all contents of Source/ThirdParty other than ANGLE
When ANGLE was moved to Source/ThirdParty as part of the merge to WebKit
r75315, it was the only item in that directory, so we took all of
Source/ThirdParty.

Since then, other items have been added upstream to Source/ThirdParty
which we don't need or want in our tree, eg
http://trac.webkit.org/changeset/81709/trunk/Source/ThirdParty/gyp

With this change we remove all contents of Source/ThirdParty other than
ANGLE. Future merges will merge only ANGLE from this directory.

Change-Id: I1eac0e2824de7f78610d0439e30dc6ba51b1d32d