Fix the crash caused by recent webkit RenderLayerCompositor change
We used to rely on the fact that RenderLayerBacking is the only subclass of
GraphicsLayerClient, but now the RenderLayerCompositor also inherit it and
is used for iframe. That cause us to get invalid pointer in GraphicsLayerAndroid.
The solution here is to disable the new webkit approach b/c we have our own
iframe support already.

bug:4592195

Change-Id: Ia786233907d83028fdc6865c28fbbd5291564ad0