Cleanup Skia related rendering code for raster rendering.
In the existing code we created a sharedBitmap for BaseTiles and
never used it instead opting for createing bitmaps on the stack to
enable partial invalidation. This CL removes the sharedBitmap concept
and instead creates all bitmaps on the stack.

This means that BackedDoubleBufferedTexture no longer needs to have
a bitmap member, but instead simply takes the bitmap from the caller
and uploads it to the appropriate texture. To make this upload clean
we now pass the bitmap via const references instead of pointers.

Change-Id: Ie218c4b4564e5574ca6e404d4857904ab41a3a5c