Remove old scalar conversions
The webkit merge from CL 112610 introduced an api change in
setrectForUnderline, changing two params from int to float. The body
of the function, however, was still performing int to float
conversions, resulting in a debug assert within skia and conversion
results forced to zero. This CL removes the outdated conversion.

Change-Id: I0d5fae39552c3e6801b83e6c4363dc063a01b165