Add Harfbuzz support for system fallback fonts
When Skia resorts to fallback fonts for non-latin characters, it
retains the font ID of originally requested font, concealing the
fact that a fallback was invoked. Harfbuzz needs to know the fallback
font ID to read the correct GSUB/GPOS tables from the fallback font
file so that these complex languages can be properly rendered.

This change uses the script recognized by Harfbuzz as a means to
surmise the fallback font that would be used by Skia and then injects
that font file as a replacement for the originally requested font.

Bug: 5087744
Change-Id: I3a3469bcd589ee796c9b5a828fb47d40ecb38738