Fix Gmail blank page
Gmail is updating the tilePages in different scale factor.
One of the beginning scale factor happens to be the same as the initial value 1.0,
then the code will consider the page is prepared which is not true, and we miss
the necessary page update.

Changing the initial value will make sure that we know the current page is not
really prepared yet. And we can update accordingly.

bug:5134126
Change-Id: Ieb4d01513fc4fbe9d996ab62ad5eb43bb5aace2e