Add Harfbuzz support for fallback fonts (redux)
When Skia resorts to fallback fonts for non-latin characters, it
retains the font ID of originally requested font, concealing the
fact that a fallback was invoked. Harfbuzz needs to know the fallback
font ID to read the correct GSUB/GPOS tables from the fallback font
file so that these complex languages can be properly rendered.

This change uses the script recognized by Harfbuzz as a means to
surmise the fallback font that would be used by Skia and then injects
that font file as a replacement for the originally requested font.

Fallback fonts are loaded once and reused multiple times. An earlier
implementation did not cache fonts in this way, causing large pages
to run out of files for mmap streaming.

Bug: 5087744
Change-Id: I61222829209d9bd6df465d37923ccff3681c2c54