Check for null selection in modifySelection
MoveSelection and supporting routines have appeared in one-off crash
reports for monkey runs. It's not yet clear whether or not these
conditions are human-reproducible, but returning an empty string when
a null selection is unexpectedly encountered in modifySelection()
should make the release build somewhat safer, while asserting in the
debug build.

Bug: 5244036
Change-Id: I4d492286cebf9884da0c675aa71175b8aa39f4b9