Limit the number of buckets in a PictureSet
Added MAX_BUCKET_COUNT_X and _Y so that absurdly large pages, such as
those in LayoutTests, don't run out of memory due to millions of
buckets.

Renamed checkDimensions() to setDimensions() and made inval an
optional parameter so that it can be used more generally within
PictureSet to adjust settings.

Bug: 5300146
Change-Id: Ic6c2dc97085ef38d5782aa005b71a6797d165e24