Add temporary work-around to allow loading local data from Documents with substituted data
This is a temporary workaround for a WebCore bug which is awaiting resolution.
See https://bugs.webkit.org/show_bug.cgi?id=68711 and this bug for details.

This change adds a new property to Frame to track when its Document pointer is
up-to-date. This is is used only when setting up the securty context for a
Document with substituted data and allows us to skip the buggy check in
Document::loader().

Bug: 5188895
Change-Id: I5d7d2cca83fa0c6db084d505f5b48207046a9cd0