Fallback handle if vertical glyphs are missing.
Bug 5094208 : browser VerticalWritingMode support.
Some punctuation has to change its glyph in vertical writing mode.
For example, dash in vertical text is a vertical bar. In
current Android system fonts and fallback fonts, none of them have the
substitute tables for vertical writing mode.  We call a conversion
function that converts the punctuation to the Unicode Vertical Form
(0xFE10 - 0xFE19) or to the closest glyphs as alternative when
vertical substitute tables are absent in fonts. Otherwise, we use
Harfbuzz to do the vertical glyph shapping.

Change-Id: If1d31bb72aaaba45ed50bbb75d6246cfa30e3428