Modify BrowserFrame.requestClientCert() to take the host and port as a String
Currently we get the host and port from Chromium as a std::string, convert to a
jbyte array to pass over JNI, then convert to String. It's simpler to convert
directly to jstring and to pass that over JNI.

This also fixes the fact that WebFrame::requestClientCert() fails to call
ReleaseByteArrayElements().

Requires https://android-git.corp.google.com/g/141238
in frameworks/base.

Bug: 5442710

Change-Id: If81fdef338ecdf2debeece0ebb9926bd261ae219