Fix uses of GetByteArrayElements() in WebCoreFrameBridge
Currently, we call GetByteArrayElements() to get a pointer to the data, then
modify it, then fail to call ReleaseByteArrayElements. This means that if
GetByteArrayElements() happens to take a copy of the data, we'll never sync the
modified data back to the JNI object.

This change updates the code to use SetByteArrayRegion() which handles
releasing any copies that are made by JNI.

Also use ScopedLocalRef where possible.

Bug: 5442710
Change-Id: Ic8691eb18cac746ee2a2086017ffb9b5558755bd