Fix crash in ImageBufferAndroid.
In the case that we do not have sufficient free
memory to create a copy of a bitmap, we must return
an Image wrapping an empty bitmap, rather than
null itself.

This handling is particular (but not exclusive) to
Android, where we cannot pass a null NativeImagePtr
like some other platforms do (as we need it to be
valid in the BitmapImage constructor).

Bug:5409973
Change-Id: I888cf1db73a9176db2a3ea2721ca0e258d9c7773