Re-enable animations on the UI thread
Using webkit to compute animations is still slow in some cases.

When animating large elements, we seems to sometimes bogs the GPU,
which then makes the UI takes longer to render a frame. This in turn
slow the rate at which we can call into webkit (to update the position
of the animated layers), which results in perceived stuttering.

We previously had an implementation of CSS animations that could run
fully on the UI thread, without having to call back into webkit.
We turned it off because there was still some glitches, and calling
into webkit seemed to work well enough -- but as we can see, even
if that's the case in general, edge cases still benefit from running
the animations outside of webkit.

The CL fixes the remaining glitches we had (mostly, it was the
non support of fillMode) and re-enable our CSS animations implementation.

bug:5297559

Change-Id: I1f00bc060a76c9dfd55bd6d8ae85d5d6da68ddb5