Doesn't lock the entire drawing...
This caused us to potentially waiting on the completion of a tile
painting. On sites where skia struggle with, this had a large impact on
scrolling performances.

This is only part one of the solution -- we need to also get rid of the
global button lock.

We also lower the priority of the texture gneeration thread.

bug:5558699

Change-Id: I251354d6cabb2fd7ebfd665c30eff7fe90c3d316