Fix font size handling for harfbuzz hack
Complex languages (Arabic, Thai, etc.) currently need to override the
base font (e.g. Roboto)  with an explicitly loaded copy of the right
fallback font (e.g. DroidNaskh). To keep from running out of file
descriptors, a caching mechanism was put in place, but the cached
data contained the font size, so the first size seen on the page was
applied to all subsequent renderings of that font, regardless of the
size requested. Size is now also considered in the cache.

Bug: 5509443
Change-Id: I616954980122e89da14ce86efe99d5e4343a40c2