Fix CTS regressions from merge to r74534.
WebKit change http://trac.webkit.org/changeset/73385 removed
overflow dimensions from the calculation of layer dimensions.
This means that the optimisation inside the nav cache to skip
building a frame cache when one of the layer dimensions is 0
is no longer valid (as the layer may contain only overflow
content).

Bug: 5572431

Change-Id: I3c973ad1ed57b07217d19f5d4fba7bfe522a2165