Promote -ve z-index fixed/scrollable elements to layers
We normally do not necessarily promote fixed or scrollable
elements to compositing layers to try and reduce the number
of layers that get created. However in the case that such
an element appears in the negative z-order, we must do
the promotion to ensure that the layer is painted at the
correct z-index in the hierarchy (i.e not too high - sites
like to use a fixed image with z-index:-1 to implement
a fixed background -- without this patch we end up painting
the background over the main page content).

Bug: 5152567

Change-Id: I36bfe14660a33d45ad54552ae41b9d7157f8dabf