Don't deep copy animations, prepare animations on both trees
 Bug: 5699085

Previously, animations were copied along with layer trees from the webkit to UI
thread. This changes that to instead use a single object per animation with
refcounting so that animations are kept consistent.

Additionally, animations are now run on both the painting and drawing trees in
the tree manager so that animated content clipped correctly during a tree swap.

Change-Id: I79f0c0e47b717f9fdddf303eb7ec29efc4950aaf