Account for dirty tiles doubly in texture counting
bug:5704511

Dirty tiles use two textures - one for display, and one for painting.

This fixes an issue where the number of textures requested did not meet the
needs of a double buffered tree with most content invalidated. The webview would
give up drawing, and thus not paint the invalidated portions of the layers.

Change-Id: Icd9b9e0b53cf82415074dac4338a8c48de880364